+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Command to eval a line (or a selection) of the active JavaScript document.
 - Command to eval a block of the active JavaScript document.
 - Hydra is exposed globally as `_hydra` (for community scripts).
+- P5 wrapper from Hydra's web editor.
 
 ## [0.1.0] - 2024-04-09